pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quantlab"
version = "0.1.0"
description = "Quantitative Trading Research Platform - Portfolio management and multi-source options analysis"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "QuantLab Team"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Office/Business :: Financial :: Investment",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "click>=8.0.0",
    "duckdb>=0.9.0",
    "pyyaml>=6.0",
    "tabulate>=0.9.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "polygon-api-client>=1.12.0",
    "yfinance>=0.2.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
quantlab = "quantlab.cli.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["quantlab"]

[tool.black]
line-length = 120
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 120
target-version = "py38"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
]

[tool.pytest.ini_options]
markers = [
    "integration: Integration tests using real components (no mocking)",
]
